fix(core): Add cache configuration settings
on the configuration we added a new flag that will tell whether we raise
Invalid in case that the cache marks some data as expired.
This is useful (and this commit contains the change) to not raise Invalid
in case that a strategy tries to return stored data with Store operation.
When we use Store operation we don't really care about the expiration of
the stored data

fixes: #47
Signed-off-by: Marc Moreno Ferrer <ignis.fatue@gmail.com>